格式化

+++ unix/apue3e/filedir/unlink.c
@@ -3,10 +3,8 @@
 
 int main(void)
 {
-    if (open("tempfile", O_RDWR) < 0)
-        err_sys("open error");
-    if (unlink("tempfile") < 0)
-        err_sys("unlink error");
+    if (open("tempfile", O_RDWR) < 0) err_sys("open error");
+    if (unlink("tempfile") < 0) err_sys("unlink error");
     printf("file unlinked\n");
     sleep(15);
     printf("done\n");
